+++ CHANGELOG.rst
@@ -1,10 +1,15 @@
 Changelog
 *********
 
+`Unreleased`_ (YYYY-MM-DD)
+--------------------------
+
+- Event callback arguments now always need to provide a `data` argument (major)
+
 `3.0.1`_ (2019-01-02)
 ---------------------
 
-- Fix forward the `timeout` close code as an `int` to event callbacks.
+- Fix forward the `timeout` close code as an `int` to event callbacks
 
 `3.0.0`_ (2018-12-18)
 ---------------------
@@ -39,13 +44,13 @@ support the `disconnected` message before upgrading.
 `1.0.2`_ (2017-11-15)
 ---------------------
 
-- Fix do not accept unencrypted 'client-auth' messages from the initiator.
+- Fix do not accept unencrypted 'client-auth' messages from the initiator
 
 `1.0.1`_ (2017-07-25)
 ---------------------
 
 - Fix to handle new `libnacl <https://github.com/saltstack/libnacl/pull/91>`_
-  exceptions.
+  exceptions
 
 `1.0.0`_ (2017-03-24)
 ---------------------
@@ -55,6 +60,7 @@ support the `disconnected` message before upgrading.
 
 .. _SaltyRTC 1.0 Protocol: https://github.com/saltyrtc/saltyrtc-meta/blob/protocol-1.0/Protocol.md
 
+.. _Unreleased: https://github.com/saltyrtc/saltyrtc-server-python/compare/v3.0.1...HEAD
 .. _3.0.1: https://github.com/saltyrtc/saltyrtc-server-python/compare/v3.0.0...v3.0.1
 .. _3.0.0: https://github.com/saltyrtc/saltyrtc-server-python/compare/v2.0.1...v3.0.0
 .. _2.0.1: https://github.com/saltyrtc/saltyrtc-server-python/compare/v2.0.0...v2.0.1